Se crean servicios de actualizar y eliminar informacion

+++ src/main/java/co/proyectoGrado/proyectoGrado/domain/model/CursoContenido.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class CursoContenido {
 
     @JsonProperty
-    private Long idCursoContenido;
+    private Integer idCursoContenido;
     @JsonProperty
     private int idCategoriaContenido;
     @JsonProperty
